# remove memoize hashRoot from `SszSuperNode` and `LazyBranchNode`