## PR Description
Substitutes the memoization of `hashTreeRoot` with a simple `volatile` variable.

## Documentation

- [ ] I thought about documentation and added the `doc-change-required` label to this PR if updates are required.

## Changelog

- [ ] I thought about adding a changelog entry, and added one if I deemed necessary.