@Nashatyrev 2022-03-29
I mean, removing `volatile` from one side increases possibility that hash would be double calculated on another thread, but from the other side it reduces synchronization overhead.