@tbenr 2022-03-29
I think the main reason of the `volatile` is to avoid threads caching "old" hashes when they have been updated by other threads. @ajsutton may be more verbose on the topic.