@ajsutton 2022-03-29
I don't think the hashes ever change (the node is replaced if either of its children are changed) but the volatile is required to ensure we have a consistent view of the `Bytes32` object. Compilers and CPUs are capable of and do reorder instructions and are free to do so as long as they respect the memory model rules. In this case, if `hashTreeRoot` wasn't `volatile` the instructions could be reordered such that the allocation to `hashTreeRoot` is done before the `Bytes32` constructor had actually completed (the `hashTreeRoot` variable is actually just a pointer to the allocated memory).  So we might see `hashTreeRoot` as not null but then when we use it the actual `bytes` field inside the `ArrayWrappingBytes32` could be `null` or only partially populated or the offset and length fields may not have been set yet etc.  The same effect could also result from memory visibility if those variables spanned multiple cache lines - we might get an up to date view of parts of them but pre-initialisation view of others.