@Nashatyrev 2022-03-30
Oh, right. Good point Adrian 👍 
Though `ArrayWrappingBytes` fields are all final and this specific case still looks safe without `volatile` it's better to be safe in case if implementation changes. 

BTW do you know if the situation you've described could be ever reproduced on x86 arch? I believe I've read that this could be reproduced on RISC arches only (which of course is also targeted by us)