@ajsutton 2022-03-30
hmm, good point on the final fields actually.  Given we're assuming we have an immutable bytes I suspect all the implementations would have that property which may make it safe to not use volatile here.  Would be interesting to know what performance difference it makes with and without.

And I think you're right that these weird cases are often risc specific - I'm not sure how they apply to ARM chipsets.  There was a really good article going into all the details from a while back that I'll have to try and dig up again.