@ajsutton 2022-03-30
It's definitely deep into the exotic java memory model kind of territory. :)