@ajsutton 2022-03-30
I think https://shipilev.net/blog/2014/safe-public-construction/ is the article I remember.