@tbenr 2022-03-31
@ajsutton @Nashatyrev 

changes applied to:
SimpleLeafNode SszSuperNode SimpleBranchNode  LazyBranchNode

Without volatile: `private volatile Bytes32 cachedHash;`
SSZBenchmark.BeaconStateSerialization  thrpt   25  46247.386 ± 952.119  ops/s
SSZBenchmark.BeaconStateSerialization  thrpt   25  47031.849 ± 643.049  ops/s
SSZBenchmark.BeaconStateSerialization  thrpt   25  47086.389 ± 635.026  ops/s

With Volatile: `private Bytes32 cachedHash;`
SSZBenchmark.BeaconStateSerialization  thrpt   25  46989.661 ± 733.422  ops/s
SSZBenchmark.BeaconStateSerialization  thrpt   25  46507.963 ± 986.361  ops/s
SSZBenchmark.BeaconStateSerialization  thrpt   25  46931.979 ± 782.850  ops/s

no measurable differences in my setup